Revisions to the ACC example: Removed movingTarget
